### 2018-11-21 #264 @kshtzgupta1
Submission: ramlegacy
`ramlegacy` helps users access the excel version of the RAM Legacy Stock Assessment Database from www.ramlegacy.org. Data is freely available from the databse website, but downloading and reading in the data by hand can be time-consuming. `ramlegacy` is capable of downloading, reading and caching all the available versions of the database.

```
Package: ramlegacy
Title: Download and Read RAM Legacy Stock Assessment DataBase
Version: 0.1.0
    c(person(given = "Kshitiz",
             family = "Gupta",
             email = "kshtzgupta1@berkeley.edu"),
      person(given = "Carl",
             family = "Boettiger",
             role = c("aut", "cph"),
             email = "cboettig@gmail.com", comment="http://orcid.org/0000-0002-1642-628X"))
Description: Contains functions to download, cache and read in Excel version of
    the RAM Legacy Stock Assessment Data Base, an online compilation of
    stock assessment results for commercially exploited marine populations
    from around the world. More information about the database can be
    found at <http://ramlegacy.org/>.
    cli (>= 1.0.0),
    crayon (>= 1.3.4),
    rappdirs (>= 0.3.1),
    readxl (>= 1.1.0)
    covr,
    httptest,
    knitr
X-schema.org-keywords: ramlegacy, marine, database, assessment, RAM, Legacy, download, read, cache, ropensci

```

 www.github.com/kshtzgupta1/ramlegacy

**data retrieval**: `ramlegacy` downloads and reads in multiple versions of the excel version of the RAM Legacy Stock Assessment Database from the database's [website.](http://www.ramlegacy.org/)

Any one who needs access to RAM Legacy Stock Assessment Database: Fisheries Biologists, Conservationists, Students, Teachers, etc.

Sean Anderson has a namesake package not published to CRAN and it appears to be a stalled project on GitHub (last updated 9 months ago). However, unlike this package which supports downloading and reading in the Excel version of the database, Sean Anderson's project downloads the Microsoft Access copy of the database and converts it to a local sqlite3 database.

There is also `RAMlegacyr`, an older package last updated in 2015. Similar to Sean Anderson's project, the package seems to be an R interface only for the Microsoft Access version of the RAM Legacy Stock Assessment Database and provides a set of functions using RPostgreSQL to connect to the database.

Please note that running R CMD check on ramlegacy may result in the following warning  if the user has not yet downloaded a version of the RAM Legacy Stock Assessment Database.

`No version of the database has yet been downloaded. Use function download_ramlegacy() to download a version now.`

 Also note that R CMD check on this package will result in the following error in the testing suite if the user is not online when checking the package.

`Error: Could not connect to the internet. Please check your connection settings and try again. `
