### 2018-12-17 #264 @geanders
- [x] **License:** The package has a CRAN or OSI accepted license (MIT + LICENSE)

---

@kshtzgupta1: Thanks for submitting a very interesting package. This is the first time I'm serving as an editor for ROpenSci, so I will be learning some parts of the editorial process as I edit your submission. Please let me know if you have any questions throughout.

I have a bit of feedback based on an initial look at the submission and some initial editorial checks.

## Submission summary:

In your submission summary, could you please:

- [ ] Add a hyperlink to the RAM Legacy Stock Assessment Database where the current summary has "[website.]"

## Checks from `devtools::spell_check()`:

There seem to be two legitimate spelling errors in the package code:

```
  WORD            FOUND IN
Assesment       load_ramlegacy.Rd:18
availabl        ram_dir.Rd:11
```

- [ ] “Assesment” to “Assessment” in “load_ramlegacy.Rd”
- [ ] “availabl” to “available” in “ram_dir.Rd”

## Checks from `covr::package_coverage()`

```
ramlegacy Coverage: 79.33%
R/zzz.R: 23.33%
R/download_ramlegacy.R: 83.02%
R/utils.R: 88.41%
R/load_ramlegacy.R: 88.46%
R/read_ramlegacy.R: 100.00%
R/style.R: 100.00%
```

## Checks from `goodpractice::gp()`

```
── GP ramlegacy ──────────────

    R/download_ramlegacy.R:58:NA
    R/download_ramlegacy.R:59:NA
    R/download_ramlegacy.R:69:NA
    R/download_ramlegacy.R:70:NA
    R/download_ramlegacy.R:71:NA

    R/download_ramlegacy.R:106:1
```

- [ ] Check code line length on the line highlighted by this check

## Checks from `devtools::check()`

- [ ] This check brings up an initial warning about “mismatched braces or quotes” in
R/utils.R#40, I believe in the sentence “This is also the location from where
\code{\link{load_ramlegacy} loads the database from.” Please add the closing brace
within that sentence.

## Other

- [ ] The first of the two “build passing” badges on the project’s GitHub page does not seem to lead
to an active Travis CI project. I suggest removing this first badge to prevent confusion.
- [ ] In the "Details" section, paste under the first item a note that running R CMD check on this package may result in a warning if the user has not yet downloaded a version of the RAM Legacy Stock Assessment Database, as well as copy in the warning that results in this case: "No version of the database has yet been downloaded. Use function download_ramlegacy() to download a version now."
- [ ] Also in the "Details" section, add a note that running R CMD check on this package will result in errors in the testing suite if the user is not online when checking the package.

In the meantime, I will begin the process of assigning reviewers.

Due date: TBD
