### 2018-12-19 6717c6d
fixed typos and long line
