+++ R/utils.R
@@ -38,7 +38,7 @@ check_path <- function(path) {
 #' @name ram_dir
 #' @description Provides the download location for \code{\link{download_ramlegacy}}
 #'  in an  OS independent manner. This is also the location from where
-#'  \code{\link{load_ramlegacy} loads the database from.
+#'  \code{\link{load_ramlegacy}} loads the database from.
 #' @param vers character, version number of the database. As of November 2018,
 #' the availabl versions are "1.0", "2.0", "2.5", "3.0" and "4.3". If version
 #' is not specified the \code{ram_dir()} returns the path to the rappdirs directory.
