+++ README.md
@@ -3,7 +3,7 @@
 ramlegacy
 =========
 
-[![Project Status: WIP – Initial development is in progress, but there has not yet been a stable, usable release suitable for the public.](https://www.repostatus.org/badges/latest/wip.svg)](https://www.repostatus.org/#wip) [![Travis build status](https://travis-ci.com/kshtzgupta1/ramlegacy.svg?branch=master)](https://travis-ci.org/kshtzgupta1/ramlegacy) [![Coverage status](https://codecov.io/gh/kshtzgupta1/ramlegacy/branch/master/graph/badge.svg)](https://codecov.io/github/kshtzgupta1/ramlegacy?branch=master) [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/kshtzgupta1/ramlegacy?branch=master&svg=true)](https://ci.appveyor.com/project/kshtzgupta1/ramlegacy)
+[![Travis Build Status](https://travis-ci.com/kshtzgupta1/ramlegacy.svg?branch=master)](https://travis-ci.com/kshtzgupta1/ramlegacy) [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/kshtzgupta1/ramlegacy?branch=master&svg=true)](https://ci.appveyor.com/project/kshtzgupta1/ramlegacy) [![Coverage status](https://codecov.io/gh/kshtzgupta1/ramlegacy/branch/master/graph/badge.svg)](https://codecov.io/github/kshtzgupta1/ramlegacy?branch=master) [![Project Status: WIP – Initial development is in progress, but there has not yet been a stable, usable release suitable for the public.](https://www.repostatus.org/badges/latest/wip.svg)](https://www.repostatus.org/#wip)
 
 -   **Authors**: Kshitiz Gupta, [Carl Boettiger](http://www.carlboettiger.info/)
 -   **License**: [MIT](http://opensource.org/licenses/MIT)
@@ -28,13 +28,13 @@ You can install the development version from [GitHub](https://github.com/kshtzgu
 ``` r
 install.packages("devtools")
 library(devtools)
-install_github("kshtzgupta1/ramlegacy")
+install_github("kshtzgupta1/ramlegacy", build_vignettes)
 ```
 
 Usage
 -----
 
-Please see the ramlegacy vignette for more detailed examples and additional package functionality.
+Please see the ramlegacy vignette for more detailed examples and additional package functionality. The vignette can be viewed by calling `vignette(package = "ramlegacy")`
 
 Start by loading the package using `library`.
 
