### 2018-12-19 d37f7f1
fix mistake in README
