+++ README.Rmd
@@ -89,7 +89,7 @@ After the specified version of the database has been downloaded through `downloa
 # load version 3.0
 load_ramlegacy(version = "3.0")
 
-# loads current latest version 4.3
+# loads the latest version (currently 4.3)
 load_ramlegacy()
 ```
 
@@ -97,10 +97,10 @@ load_ramlegacy()
 To view the exact path where a certain version of the database was downloaded and cached by `download_ramlegacy` you can run `ram_dir(vers = 'version')`, specifying the version number inside the function call:
 ```{r, ram_dir_example1, eval = F, echo = T}
 # downloads version 2.5
-download_ramlegacy(version = '2.5')
+download_ramlegacy(version = "2.5")
 
-# view the location where version 2.5 of the database was downloaded
-ram_dir(vers = '2.5')
+# view the location where version 2.5 of the database was downloaded and cached
+ram_dir(vers = "2.5")
 ```
 
 ## Similar Projects
