### 2019-01-08 #264 @boshek
- Though this is likely to end up on CRAN, the current installation instructions don't automatically build the vignette so accessing the vignette locally fails. I'd just suggest remove the line:

  > The vignette can also be viewed by calling vignette(package = "ramlegacy")

  Until the installation instruction are for the CRAN version. Alternatively, one could install by removing *--no-build-vignettes*
  from the `build_opts` in `install_github`
  - Use quoting markdown for RAM legacy description
  - This line:
https://github.com/kshtzgupta1/ramlegacy/blob/8c26a196cc573a03574ed084a7e6c9d0b951e7a2/vignettes/ramlegacy.Rmd#L37
  I believe is a typo and should read `library(ramlegacy)`
  - >Note that this function does not support vectorization so please don't pass in a vector of version numbers
  - Consider revising the above since you are unable to do so anyways.
  - Consider adding either suggested reference or the maintainer of ramlegacy to the DESCRIPTION file in Authors@R
  - Consider adding a CODE OF CONDUCT and CONTRIBUTING file directly to the repo.

  - It does with the exception of the vignettes not being build (discussed above)

---

I think this is a great package that provides a clear set of instruction on how to download the ramlegacy database. The biggest issue in the package I see is the reliance of the `library` call to load data. The rest of the items identified below are relatively easy to address. Please let me know if anything is unclear or there is any disagreement.

- I think that you are adding too much onto the `library` call. Consider any loading of data should be left solely to the `load_ramlegacy` function. I think this constrains future development of the package as future may entail functionality outside the database. This is also outside what a user would expect from a typical `library` call. We are into stylistic territory here but I personally think objects should not be loaded with the library call.

- I wonder if `load_ramlegacy` should gain an argument where a user can specify a vector that controls which dataset are loaded. This might be a useful feature for production code, particularly a shiny app.

- Consider not even exposing a `raw_url` nor `ram_path` argument here as described here:
https://github.com/kshtzgupta1/ramlegacy/blob/8c26a196cc573a03574ed084a7e6c9d0b951e7a2/R/download_ramlegacy.R#L23-L25
If a user is never meant to change it, then perhaps the argument could be omitted abstracting that possible confusion away from the user. What is the utility of those arguments if they are never meant to change?

- I would strongly consider a NEWS.md file to track changes between versions. rOpenSci provides a great [template](https://ropensci.github.io/dev_guide/releasing.html#news)

- I think the package lacks some connection with the actual data. Right now a bunch of dataframes are loaded but a new user is expected to have prior knowledge of the ram database. I think this package could provide some path for users to learn about the data itself. Options include:
  - Use some of the documentation that is downloaded in the .zip file to clarify (perhaps via internal data) what each table means. Many of the column names aren't terribly descriptive. Somehow ramlegacy should provide a path for users to figure that out.

  - Perhaps the vignette/README could link to documentation. As it stands, the current documentation is very clear to load some tables in the global environment but this only takes the user partially up the learning curve. ramlegacy should provide users a path to take the next steps.

  - the vignette to include a very brief sample analysis where two tables are joined. A trivial example would be something like this:

``` r
library(ramlegacy)
#> * Multiple versions found including the latest one: 4.3 . Loading the latest version.
#> * Loading version 4.3 ...
#> v Version 4.3 has been successfully loaded.

TC_with_names <- merge(
  timeseries_values_views_v4.3[!is.na(timeseries_values_views_v4.3$TC), c("stockid", "TC")],
  stock_v4.3[stock_v4.3$region == "West Africa",],
  by = "stockid"
)

hist(TC_with_names$TC)
```

- It isn't clear to me why you need different versions of the database. This functionality is highlighted quite prominently but it's utility is not clear.

- RE: the alternate download location on github. I would consider if you have acknowledged and licensed this data appropriately in your assets repo. I would contact the RAM maintainers directly and seek direction on this. Fully reproducing the data might cause some concern with those maintainers.

- The docs directory,  `_pkgdown.yml` and `_config.yml` all needed to be added to `.Rbuildignore` for R CMD check to pass on my machine.

- I am wondering if this is intentional behaviour. I think this is a carryover from `readxl::read_excel` but
given the class of the data object it prints either as a dataframe or as a tibble depending on whether tibble is loaded or not. One thing to consider is even just adding tibble to Suggests for nicer printing in a vignette:
``` r
library(ramlegacy)
#> * Multiple versions found including the latest one: 4.3 . Loading the latest version.
#> * Loading version 4.3 ...
#> v Version 4.3 has been successfully loaded.

class(stock_v4.3)
#> [1] "tbl_df"     "tbl"        "data.frame"

head(stock_v4.3, 10)
#>             stockid    tsn     scientificname              commonname
#> 1       ACADRED2J3K 166774 Sebastes fasciatus         Acadian redfish
#> 2  ACADRED3LNO-UT12 166774 Sebastes fasciatus         Acadian redfish
#> 3      ACADREDGOMGB 166774 Sebastes fasciatus         Acadian redfish
#> 4        ACADREDUT3 166774 Sebastes fasciatus         Acadian redfish
#> 5        ACMACKSARG 172413     Scomber colias Argentine chub mackerel
#> 6           AFLONCH 166156    Beryx splendens               Alfonsino
#> 7            ALBAIO 172419   Thunnus alalunga           albacore tuna
#> 8           ALBAMED 172419   Thunnus alalunga           albacore tuna
#> 9          ALBANATL 172419   Thunnus alalunga           Albacore tuna
#> 10         ALBANPAC 172419   Thunnus alalunga           Albacore tuna
#>                      areaid                                    stocklong
#> 1           Canada-DFO-2J3K                    Acadian redfish NAFO-2J3K
#> 2      Canada-DFO-3LNO-UT12      Acadian redfish Units 1-2 and NAFO-3LNO
#> 3              USA-NMFS-5YZ Acadian redfish Gulf of Maine / Georges Bank
#> 4            Canada-DFO-UT3                       Acadian redfish Unit 3
#> 5       Argentina-CFP-ARG-S   Argentine chub mackerel Southern Argentina
#> 6   multinational-SPRFMO-CH                              Alfonsino Chile
#> 7     multinational-IOTC-IO                   Albacore tuna Indian Ocean
#> 8   multinational-ICCAT-MED                  Albacore tuna Mediterranean
#> 9  multinational-ICCAT-NATL                 Albacore tuna North Atlantic
#> 10   Multinational-ISC-NPAC                  Albacore tuna North Pacific
#>                     region inmyersdb myersstockid
#> 1        Canada East Coast         0         <NA>
#> 2        Canada East Coast         0         <NA>
#> 3            US East Coast         0         <NA>
#> 4        Canada East Coast         0         <NA>
#> 5            South America         0         <NA>
#> 6            South America         0         <NA>
#> 7             Indian Ocean         0         <NA>
#> 8  Mediterranean-Black Sea         0         <NA>
#> 9           Atlantic Ocean         0         <NA>
#> 10           US West Coast         0         <NA>

library(tibble)

stock_v4.3
#> # A tibble: 1,294 x 9
#>    stockid    tsn scientificname commonname areaid stocklong region
#>    <chr>    <dbl> <chr>          <chr>      <chr>  <chr>     <chr>
#>  1 ACADRE~ 166774 Sebastes fasc~ Acadian r~ Canad~ Acadian ~ Canad~
#>  2 ACADRE~ 166774 Sebastes fasc~ Acadian r~ Canad~ Acadian ~ Canad~
#>  3 ACADRE~ 166774 Sebastes fasc~ Acadian r~ USA-N~ Acadian ~ US Ea~
#>  4 ACADRE~ 166774 Sebastes fasc~ Acadian r~ Canad~ Acadian ~ Canad~
#>  5 ACMACK~ 172413 Scomber colias Argentine~ Argen~ Argentin~ South~
#>  6 AFLONCH 166156 Beryx splende~ Alfonsino  multi~ Alfonsin~ South~
#>  7 ALBAIO  172419 Thunnus alalu~ albacore ~ multi~ Albacore~ India~
#>  8 ALBAMED 172419 Thunnus alalu~ albacore ~ multi~ Albacore~ Medit~
#>  9 ALBANA~ 172419 Thunnus alalu~ Albacore ~ multi~ Albacore~ Atlan~
#> 10 ALBANP~ 172419 Thunnus alalu~ Albacore ~ Multi~ Albacore~ US We~
#> # ... with 1,284 more rows, and 2 more variables: inmyersdb <dbl>,
#> #   myersstockid <chr>
```

- I get these errors when I run `goodpractice::gp`
```
  ✖ fix this R CMD check WARNING: LaTeX errors when creating PDF
    version. This typically indicates Rd problems.
  ✖ fix this R CMD check ERROR: Re-running with no redirection of
    stdout/stderr. Hmm ... looks like a package You may want to clean
    up by 'rm -Rf
    C:/Users/salbers/AppData/Local/Temp/Rtmp4yG4PY/Rd2pdf43d827bd526'
```

- I think some slight improvement could be made by running `styler` on the package though it is marginal.

#### Possible future work
I wonder if it would be useful to engage Sean Anderson to see if he would be willing to submit a pull request with his code that converts the Access database into an sqlite database. I think there would be an appetite for the sqlite option which could exist alongside the Excel version. His process requires an additional utility outside of R but maybe a R-only solution is available.
