### 2019-01-23 bcd48ae
fix typos, revise documentation, improve styling
