+++ tests/testthat/test_utils.R
@@ -3,18 +3,22 @@ context("Testing functions in utils.R")
 ram_url <- "https://depts.washington.edu/ramlegac/wordpress/databaseVersions"
 
 test_that("net_check doesn't error behind proxy server with show_error as F", {
-httr::with_config(httr::use_proxy(url = "http://google.com", port = 1234), {
-                    expect_silent(net_check(ram_url, show_error = FALSE))
-                      expect_false(net_check(ram_url, show_error = FALSE))
-                    })
+  httr::with_config(httr::use_proxy(url = "http://google.com", port = 1234), {
+    expect_silent(net_check(ram_url, show_error = FALSE))
+    expect_false(net_check(ram_url, show_error = FALSE))
+  })
 })
 
 test_that("net_check errors behind proxy server with show_error as TRUE", {
-httr::with_config(httr::use_proxy(url = "http://google.com", port = 1234), {
-                  expect_error(net_check(ram_url, show_error = TRUE),
-                    paste("Could not connect to the internet.",
-                    "Please check your connection settings and try again."))
-                  })
+  httr::with_config(httr::use_proxy(url = "http://google.com", port = 1234), {
+    expect_error(
+      net_check(ram_url, show_error = TRUE),
+      paste(
+        "Could not connect to the internet.",
+        "Please check your connection settings and try again."
+      )
+    )
+  })
 })
 
 test_that("net_check works when no connection issues w/ show_error as TRUE", {
@@ -28,17 +32,21 @@ test_that("net_check works when no connection issues w/ show_error as FALSE", {
 })
 
 test_that("net_check doesn't error when no net connection w/ show_error as F", {
-httptest::without_internet({
-  expect_silent(net_check(ram_url, show_error = FALSE))
-  expect_false(net_check(ram_url, show_error = FALSE))
+  httptest::without_internet({
+    expect_silent(net_check(ram_url, show_error = FALSE))
+    expect_false(net_check(ram_url, show_error = FALSE))
   })
 })
 
 test_that("net_check errors if no internet connection with show_error as T", {
   httptest::without_internet({
-    expect_error(net_check(ram_url, show_error = TRUE),
-                 paste("Could not connect to the internet.",
-                       "Please check your connection settings and try again."))
+    expect_error(
+      net_check(ram_url, show_error = TRUE),
+      paste(
+        "Could not connect to the internet.",
+        "Please check your connection settings and try again."
+      )
+    )
   })
 })
 
@@ -55,9 +63,9 @@ test_that("find_latest behaves correctly", {
   })
   # test find_latest behaves as expected with proxy connection issues
   httr::with_config(httr::use_proxy(url = "http://google.com", port = 1234), {
-                    expect_silent(find_latest(ram_url))
-                    expect_equal(find_latest(ram_url), "4.3")
-                    })
+    expect_silent(find_latest(ram_url))
+    expect_equal(find_latest(ram_url), "4.3")
+  })
   # test find_latest behaves as expected when response has failed
   expect_equal(find_latest(test_url1), "4.3")
   expect_equal(find_latest(test_url3), "4.3")
@@ -101,7 +109,6 @@ test_that("write_version works as expected", {
 
   # remove the contents of the temp paths
   unlink(test_path, recursive = TRUE)
-
 })
 
 test_that("find_local behaves as expected", {
@@ -149,18 +156,30 @@ test_that("check_version_arg works with valid versions", {
 
 
 test_that("check_version_arg fails with invalid versions", {
-  expect_error(check_version_arg(1.1),
-               "Invalid version number. Available versions are")
-  expect_error(check_version_arg(1.5),
-               "Invalid version number. Available versions are")
-  expect_error(check_version_arg(2.4),
[…]
