### 2019-01-23 d63b1b5
update doc and vignette with v4.4
