+++ R/download_ramlegacy.R
@@ -26,7 +26,7 @@
 #' @examples
 #' \dontrun{
 #'
-#' # If version is not specified then current latest version (4.3)
+#' # If version is not specified then current latest version (4.4)
 #' # will be downloaded
 #' download_ramlegacy()
 #'
