+++ vignettes/ramlegacy.Rmd
@@ -22,7 +22,7 @@ From the database's [website](www.ramlegacy.org):
 
 > The RAM Legacy Stock Assessment Database is a compilation of stock assessment results for commercially exploited marine populations from around the world. It is inspired by Dr. Ransom A. Myers' original [stock-recruitment database](http://www.mathstat.dal.ca/~myers/welcome.html), which is no longer being updated."
 
-Different versions of the database are available in compressed (zipped) form on the [website](www.ramlegacy.org) in two formats: **Access** (Microsoft Office Access 2007 file format)  and **Excel** (Microsoft Excel Open XML Format Spreadsheet). As of November 2018, the available versions are 1.0, 2.0, 2.5, 3.0, 4.3.
+Different versions of the database are available in compressed (zipped) form on the [website](www.ramlegacy.org) in two formats: **Access** (Microsoft Office Access 2007 file format)  and **Excel** (Microsoft Excel Open XML Format Spreadsheet). As of January 2019, the available versions are 1.0, 2.0, 2.5, 3.0, 4.3, and 4.4.
 
 ## About the package
 The goal of `ramlegacy` is to cache the specified version/versions of the RAM Legacy Stock Assessment Excel Database and this way save time and effort spent in re-downloading different
