+++ man/load_ramlegacy.Rd
@@ -27,7 +27,7 @@ Loads all the dataframes present in the specified version of the database into
 \examples{
 \dontrun{
 # If version is not specified then current latest version
-# (4.3) will be loaded
+# (4.4) will be loaded
 load_ramlegacy()
 
 # load version 3.0
