### 2019-01-24 #264 @jafflerbach
I just wanted to follow up and it seems within the past week, the maintainers of ramlegacy.org have changed where the database is kept. It is [now hosted on Zenodo](https://sites.uw.edu/ramlegac/database/), which will require some changes to this package for accessing the data.
