### 2019-01-24 #264 @geanders
@boshek : Just to clarify, you're referring to the "DBdata.RData" and "loadDBdata (assessment data only).r" files listed [here](https://zenodo.org/record/2542919#.XEo5hi2ZN8f), right?
