### 2019-02-12 #264 @kshtzgupta1
@geanders : Apologies for replying so late. We are working to incorporate some of the changes suggested by the reviewers as well as improve the package to better deal with the migration of the latest versions to Zenodo. I will be meeting with Prof. Boettiger this Thursday to finalize those changes after which I will submit the pull request addressing each of the reviewers' comments.
