+++ R/load_ramlegacy.R
@@ -1,16 +1,16 @@
 
 #' @name load_ramlegacy
 #' @title Read-in downloaded RAM Legacy Database
-#' @description Loads all the dataframes present in the specified version of the database into
-#'  user's global environment
+#' @description Returns a list of all the dataframes present in the specified
+#'  version of the database.
 #' @param version A character vector of length 1 specifying the version number of the
 #'  database. As of January 2019, the available versions are
 #'  "1.0", "2.0", "2.5", "3.0", "4.3", and "4.4". If version argument is not specified then it
 #'  defaults to latest version (currently "4.4"). If you want to load multiple versions please
 #'  load them one at a time as passing them all at once will throw an error.
-#' @param path path to the local directory where the specified version of
+#' @param ram_path path to the local directory where the specified version of
 #' the RAM Legacy Stock Excel Assessment Database was downloaded.
-#'  This path can be viewed using calling the function \code{\link{ram_dir}} and specifying
+#'  By default this path is set to can be viewed using calling the function \code{\link{ram_dir}} and specifying
 #'  the version number inside the function call. This function \strong{does not} support
 #'  setting a user-specified path so \strong{please
 #'  do not pass} in a path argument to \code{path}.
@@ -29,61 +29,42 @@
 #' load_ramlegacy(version = "2.5")
 #' }
 #'
-load_ramlegacy <- function(version = NULL, path = NULL) {
-  ram_url <- "https://depts.washington.edu/ramlegac/wordpress/databaseVersions"
+load_ramlegacy <- function(version = NULL, ram_path = NULL) {
+  ram_url <- "https://doi.org/10.5281/zenodo.2542918"
 
   if (!is.null(version)) {
-    version <- sprintf("%.1f", as.numeric(version))
+    # ensure that the version is properly formatted
+    version <- fmt_version(version)
     check_version_arg(version)
   } else {
     version <- find_latest(ram_url)
   }
 
-  if (is.null(path)) {
-    path <- file.path(ram_dir(vers = version), paste0("v", version, ".rds"))
+  if (is.null(ram_path)) {
+
+    if (version < "4.40") {
+      rds_path <- file.path(ram_dir(vers = version), paste0("v", version, ".rds"))
+
+      } else {
+
+      rds_path <- file.path(ram_dir(vers = version), paste0("RLSADB v", version))
+      rds_path <- file.path(rds_path, "DB Files With Assessment Data")
+      rds_path <- file.path(rds_path, paste0("v", version, ".rds"))
+
+      }
+
   } else {
-    check_path(path)
-  }
 
-  notify(paste("Loading version", version, "..."))
+    check_path(ram_path)
+    rds_path <- ram_path
+
+    }
+
   # make sure version is present
-
-  if (!file.exists(path)) {
+  if (!file.exists(rds_path)) {
     stop(paste0("Version ", version, " not found locally."))
   }
 
-  # read in the list of dataframes
-  list_dataframes <- readRDS(path)
-
-  # add the version number at end of all dataframe names
-  names(list_dataframes) <- paste0(names(list_dataframes), "_v", version)
-
-  # assign all the dataframes to the global env
-  lapply(
-    seq_along(list_dataframes),
-    function(i) {
-      delayedAssign(names(list_dataframes)[i],
-        list_dataframes[[i]],
-        assign.env = .GlobalEnv
-      )
-    }
-  )
-
-  # check whether all the dataframes are now present in the global environment
-  # through creating a boolean vecor exists_vec
-  exists_vec <- unlist(lapply(
-    names(list_dataframes),
-    function(i) {
-      exists(i, envir = .GlobalEnv)
-    }
-  ))
-
-
[…]
