+++ R/download_ramlegacy.R
@@ -2,11 +2,11 @@
 #' @name download_ramlegacy
 #' @title Download RAM Legacy Excel Database
 #' @description Downloads a specified version of RAM Legacy Stock Assessment
-#'  Excel Database as an RDS object to a local directory specified by \code{\link{ram_dir}}.
+#'  Excel Database and as an RDS object to a local directory specified by \code{\link{ram_dir}}.
 #'  The function will check if the version requested already exists
 #'  and if it is then it's up to the user to download it again. The function also
-#'  supports downloading all the versions from [backup location](www.github.com/kshtzgupta1/ramlegacy-assets)
-#'  in case the database [website](www.ramlegacy.org) is down.
+#'  supports downloading all the older versions (1.0, 2.0, 2.5, 3.0, 4.3) from
+#'  [a github repo](www.github.com/kshtzgupta1/ramlegacy-assets)
 #' @param version A character vector of length 1 specifying the version number
 #'  of the database that should be downloaded. As of January 2019, the available versions are "1.0",
 #'  "2.0", "2.5", "3.0", "4.3", and "4.4". If the version argument is not specified then it defaults
@@ -19,8 +19,8 @@
 #'  by calling \code{\link{ram_dir}}. This function \strong{does not} support
 #'  setting a user-specified path so \strong{please
 #'  do not pass} in a path to \code{ram_path}.
-#' @param ram_url A string. By default it is set to the url of the ramlegacy
-#' wordpress database website. Please \strong{do not pass} in any other url to
+#' @param ram_url A string. By default it is set to the url of zenodo.
+#'  Please \strong{do not pass} in any other url to
 #'  \code{ram_url}.
 #' @export
 #' @examples
@@ -37,16 +37,21 @@
 #' download_ramlegacy(version = "4.3")
 #' }
 download_ramlegacy <- function(version = NULL, ram_path = NULL,
-ram_url = "https://depts.washington.edu/ramlegac/wordpress/databaseVersions") {
+ram_url = "https://doi.org/10.5281/zenodo.2542918") {
 
-  # Get the latest version from web
+  # user_path, a boolean flag set to FALSE by default
+  user_path <- FALSE
+
+  # According to the Zenodo website this DOI will always resolve to the latest url.
+
+  # Get the latest version from the ram_url
   latest_vers <- find_latest(ram_url)
 
-  # version argument
+  # the user has specified a version
   if (!is.null(version)) {
-
     # make sure the version argument is formatted correctly
-    version <- sprintf("%.1f", as.numeric(version))
+    version <- fmt_version(version)
+    # check that it is a valid version
     check_version_arg(version)
   } else {
     version <- latest_vers
@@ -59,11 +64,13 @@ ram_url = "https://depts.washington.edu/ramlegac/wordpress/databaseVersions") {
   } else {
     check_path(ram_path)
     vers_path <- file.path(ram_path, version)
+    user_path <- TRUE
   }
 
-  ## If there is an existing ramlegacy version
+  ## Provided that the path is not set by user if
+  # there is an existing ramlegacy version
   # ask the user what to do in interactive mode otherwise exit
-  if (version %in% find_local(ram_path, latest_vers)) {
+  if (version %in% find_local(ram_path) & !user_path) {
     if (interactive()) {
       ans <- ask_yn(
         "Version ", version, " has already been downloaded.",
@@ -86,61 +93,60 @@ ram_url = "https://depts.washington.edu/ramlegac/wordpress/databaseVersions") {
   # check internet connection and throw error if there is a connection issue
   net_check(ram_url, show_error = TRUE)
 
-  # construct url to download from
 
[…]
