+++ man/ram_dir.Rd
@@ -4,10 +4,10 @@
 \alias{ram_dir}
 \title{Output OS-independent path to the downloaded RAM Legacy database}
 \usage{
-ram_dir(vers = NULL)
+ram_dir(version = NULL)
 }
 \arguments{
-\item{vers}{character, version number of the database. As of January 2019,
+\item{version}{character, version number of the database. As of January 2019,
 the available versions are "1.0", "2.0", "2.5", "3.0", "4.3", and "4.4" . If version
 is not specified the \code{ram_dir()} returns the path to the rappdirs directory.}
 }
@@ -23,5 +23,5 @@ ram_dir()
 # Returns the path to version 4.3 subdirectory of the rappdirs directory.
 # This is the path where version 4.3 of the database is downloaded to and
 # read from.
-ram_dir(vers = "4.3")
+ram_dir(version = "4.3")
 }
