### 2019-02-19 7c6e6f0
fixed bug in find_latest()
