### 2019-02-19 81d565a
add CITATION, CODE_OF_CONDUCT, NEWS, CONTRIBUTING
