+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Type: Package
 Package: ramlegacy
-Title: Download and Read RAM Legacy Stock Assessment DataBase
+Title: Download And Read RAM Legacy Stock Assessment DataBase
 Version: 0.1.0
 Authors@R: 
     c(person(given = "Kshitiz",
@@ -12,12 +12,16 @@ Authors@R:
              role = c("aut", "cph"),
              email = "cboettig@gmail.com", comment="http://orcid.org/0000-0002-1642-628X"),
       person(given = "RAM Legacy Stock Assessment Database",
-      role = c("ctb"))       )
+      role = c("ctb", "dtc")))
 Description: Contains functions to download, cache and read in Excel version of
     the RAM Legacy Stock Assessment Data Base, an online compilation of
     stock assessment results for commercially exploited marine populations
     from around the world. More information about the database can be
     found at <http://ramlegacy.org/>.
+    Ricard, D., Minto, C., Jensen, O.P. and Baum, J.K. (2012) Evaluating the knowledge
+    base and status of commercially exploited marine species with the RAM Legacy Stock
+    Assessment Database. Fish and Fisheries 13 (4) 380-398.
+    <doi:10.1111/j.1467-2979.2011.00435.x.>
 License: MIT + file LICENSE
 URL: https://github.com/kshtzgupta1/ramlegacy
 BugReports: https://github.com/kshtzgupta1/ramlegacy/issues
@@ -40,4 +44,4 @@ VignetteBuilder:
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.1.1
-X-schema.org-keywords: ramlegacy, marine, database, assessment, RAM, Legacy, download, read, cache, ropensci
+X-schema.org-keywords: ramlegacy, fish, marine, database, assessment, RAM, Legacy, download, read, cache, ropensci
