### 2019-02-19 493eec4
fix check warnings
