+++ R/utils.R
@@ -53,7 +53,7 @@ check_path <- function(path) {
 #' @description Provides the download location for \code{\link{download_ramlegacy}}
 #'  in an  OS independent manner. This is also the location from where
 #'  \code{\link{load_ramlegacy}} loads the database from.
-#' @param version character, version number of the database. As of January 2019,
+#' @param vers character, version number of the database. As of January 2019,
 #' the available versions are "1.0", "2.0", "2.5", "3.0", "4.3", and "4.4" . If version
 #' is not specified the \code{ram_dir()} returns the path to the rappdirs directory.
 #' @export
@@ -64,7 +64,7 @@ check_path <- function(path) {
 #' # Returns the path to version 4.3 subdirectory of the rappdirs directory.
 #' # This is the path where version 4.3 of the database is downloaded to and
 #' # read from.
-#' ram_dir(version = "4.3")
+#' ram_dir(vers = "4.3")
 ram_dir <- function(vers = NULL) {
   if (!is.null(vers)) {
     version <- fmt_version(vers)
