+++ man/ramlegacy-package.Rd
@@ -4,14 +4,12 @@
 \name{ramlegacy-package}
 \alias{ramlegacy}
 \alias{ramlegacy-package}
-\title{ramlegacy: Download and Read RAM Legacy Stock Assessment DataBase}
+\title{ramlegacy: Download And Read RAM Legacy Stock Assessment DataBase}
 \description{
-Contains functions to download, cache and read in Excel version of
-    the RAM Legacy Stock Assessment Data Base, an online compilation of
-    stock assessment results for commercially exploited marine populations
-    from around the world. More information about the database can be
-    found at <http://ramlegacy.org/>.
-    Ricard, D., Minto, C., Jensen, O.P. and Baum, J.K. (2012) Evaluating the knowledge
+Contains functions to download, cache and read in Excel version of the
+    RAM Legacy Stock Assessment Data Base, an online compilation of stock assessment
+    results for commercially exploited marine populations from around the world. More information about the database can be
+    found at <http://ramlegacy.org/>. Ricard, D., Minto, C., Jensen, O.P. and Baum, J.K. (2012) Evaluating the knowledge
     base and status of commercially exploited marine species with the RAM Legacy Stock
     Assessment Database. Fish and Fisheries 13 (4) 380-398.
     <doi:10.1111/j.1467-2979.2011.00435.x.>
