### 2019-02-19 8f28313
fixed DESCRIPTION and debugged load_ramlegacy
