+++ man/ramlegacy-package.Rd
@@ -8,11 +8,11 @@
 \description{
 Contains functions to download, cache and read in Excel version of the
     RAM Legacy Stock Assessment Data Base, an online compilation of stock assessment
-    results for commercially exploited marine populations from around the world. More information about the database can be
-    found at <http://ramlegacy.org/>. Ricard, D., Minto, C., Jensen, O.P. and Baum, J.K. (2012) Evaluating the knowledge
-    base and status of commercially exploited marine species with the RAM Legacy Stock
-    Assessment Database. Fish and Fisheries 13 (4) 380-398.
-    <doi:10.1111/j.1467-2979.2011.00435.x.>
+    results for commercially exploited marine populations from around the world. More
+    information about the database can be found at <http://ramlegacy.org/>. Ricard, D.,
+    Minto, C., Jensen, O.P. and Baum, J.K. (2012) Evaluating the knowledge base and
+    status of commercially exploited marine species with the RAM Legacy Stock Assessment
+    Database. Fish and Fisheries 13 (4) 380-398. <doi:10.1111/j.1467-2979.2011.00435.x.>.
 }
 \references{
 Ricard, D., Minto, C., Jensen, O.P. and Baum, J.K. (2012)

[…]
