### 2019-02-19 #264 @kshtzgupta1
@geanders @boshek @jafflerbach Thanks for these very helpful and detailed reviews! I believe Prof. Boettiger and I have addressed most of these issues now. Among other things, we have removed the loading behavior of `library(ramlegacy)` leaving the loading solely to `load_ramlegacy ` which instead of assigning all the tables to global environment now just returns a list of the tables. Additionally, if the user wants specific tables from a version of the database `load_ramlegacy` now supports that behavior. Finally, while the default is to download and read in from the rappdirs directory the package now also supports downloading to a location chosen by the user and reading from that location. I've made these edits in a separate branch and created a PR, [kshtzgupta1/ramlegacy#1](https://github.com/kshtzgupta1/ramlegacy/pull/1) which I will leave un-merged for now, hopefully that provides an easy way to see the differences. I've also summarized my replies to each of your point-by-point reviews (with a few clarifying questions and comments) in the two issues linked in that PR.
