### 2019-02-26 #264 @kshtzgupta1
@jafflerbach  Thank you for a detailed review! We have made many changes following your suggestions. I will update the README and vignette to reflect those changes after they are approved.

>  Vignette(s) demonstrating major functionality that runs successfully locally
The Vignette is good except it does not explain how to install the package. You can only find installation instructions in the README

Fixed!

> * When running `goodpractice::gp()` I got the following output:

```
  ✖ write unit tests for all functions, and all package code in general. 69% of code lines are covered by test
    cases.

    R/download_ramlegacy.R:53:NA
    R/download_ramlegacy.R:58:NA
    R/download_ramlegacy.R:59:NA
    R/download_ramlegacy.R:68:NA
    R/download_ramlegacy.R:69:NA
    ... and 58 more lines

  ✖ fix this R CMD check WARNING: LaTeX errors when creating PDF version. This typically indicates Rd
    problems.
```

We couldn't reproduce that warning on our machines. I think it might be specific to your machine and may be occurring because you don't have Tex installed.

> That being said, I strongly encourage the developers of this package to work with Sean and the maintainers of the RAM legacy database to get this package functional.

I believe Prof. Boettiger has reached out to Sean regarding that.

> I first tried `download_ramlegacy()` when the site was down (1/17/2019). I downloaded from the backup location on github. This worked well. A couple of hours later, the website was backup and when I tried to download the _only_ version now available on the site (4.4) I got an error message

```r
> download_ramlegacy(version = "4.4")
Error: Invalid version number. Available versions are 1.0, 2.0, 2.5, 3.0, 4.3,
```

This should be resolved now!

>I can forsee some user issues with the `ram_url()` argument in `download_ramlegacy()`:

```r
download_ramlegacy(version = NULL, ram_path = NULL,
  ram_url = "https://depts.washington.edu/ramlegac/wordpress/databaseVersions")
```

> First, it looks like there is a typo (`ramlegac` instead of `ramlegacy`) so if the `download_ramlegacy()` function isn't acting as the user would expect (whether it's their fault or not) they might see the URL in the Help document and manually try to fix the typo. This may be an actual typo, I don't know. Either way, if you don't want the user to ever touch it I would hardcode that `ram_url` into the function itself rather than having it as an argument.

It was actually `ramlegac` in the url. But that is no longer relevant since we are now passing in the zenodo doi url to `ram_url`. Having `ram_url` as an argument helps us test for intended behavior of `download_ramlegacy`  in face of internet connection issues and network problems.

> Also the `path` argument in `load_ramlegacy()` is again one that should not be touched by the user. I tried to mess with this path argument and when I entered an incorrect path I received this:

```r
> load_ramlegacy(version = 4.3, path = "~/github")
★ Loading version 4.3 ...
Error in readRDS(path) : error reading from connection
In addition: Warning message:
In readRDS(path) : error reading the file
```

> I suggest changing the function so that if something is entered in as a path argument, it returns a message reminding the user to not change the path argument. Or, remove this as an argument entirely.

Although the package downloads and caches the database in the user's rappdirs directory by default we have now decided that having `ram_path` as an argument is useful in `download_ramlegacy` and `load_ramlegacy` because it allows the user to download the database to a location of their choice and read from that location if they wish to do so. I think we want to give the user that flexibility.
