+++ R/load_ramlegacy.R
@@ -8,7 +8,7 @@
 #'  "4.40", "4.41" and "4.44". If version argument is not specified then it defaults to latest version
 #'  (currently "4.44"). If you want to load multiple versions please
 #'  load them one at a time as passing them all at once will throw an error.
-#' @param dfs A character vector specifying the names of the dataframes to load from a certain version
+#' @param tables A character vector specifying the names of the dataframes to load from a certain version
 #' @param ram_path path to the local directory where the specified version of
 #' the RAM Legacy Stock Excel Assessment Database was downloaded.
 #' By default this path is set to within the rapddirs directory and can be viewed using calling the function
@@ -31,7 +31,7 @@
 #' load_ramlegacy(version = "2.5")
 #' }
 #'
-load_ramlegacy <- function(version = NULL, dfs = NULL, ram_path = NULL) {
+load_ramlegacy <- function(version = NULL, tables = NULL, ram_path = NULL) {
   ram_url <- "https://doi.org/10.5281/zenodo.2542918"
 
   if (!is.null(version)) {
