### 2019-03-02 #264 @cboettig
Hi Sam @boshek , thanks for the excellent review and the kinds words! I don't mean to jump in here but just wanted to add a little context to the decision about accessing older versions.  I agree that it would be much more desirable to have all versions on Zenodo, and based on our discussions with them, I think the RAM Legacy team will eventually post those, but it is hard to know exactly when.  Meanwhile, I do believe access to the old versions is critical, all the more so for them not being available anywhere else now.  Among other reasons, there are dozens of high profile papers based on these older versions; just today there is yet [another](https://doi.org/10.1126/science.aau1758) appearing in *Science* that uses the version 3 data.
