### 2019-03-06 #264 @boshek
Great context @cboettig . This rationale makes perfect sense.
