### 2019-03-08 #264 @cboettig
Also just a note that we did just hear back from Sean Anderson who was very gracious and positive about the package.  He gave us a few suggestions (data tables no longer append the version name on the table, which was no longer necesary anyway with new approach of explicitly calling `load_ramlegacy()` with a version number instead of trying to load automatically on `library()`.  (Also note that as Sean's package no longer works in it's current state, as it still points to the Wordpress URLs which are now defunct and would need to be replaced with the Zenodo DOIs)

Thanks Sam,  Jamie & Brooke for your feedback so far!  @kshtzgupta1 summarizes the edits in reply to Jamie [above](https://github.com/ropensci/software-review/issues/264#issuecomment-467565078), these are waiting on an [PR from an ropensci-review branch](https://github.com/kshtzgupta1/ramlegacy/pull/1) in the package meanwhile.
