### 2019-03-15 #264 @geanders
Also, since you are planning to submit to CRAN, you may find this list of [CRAN gotchas](https://ropensci.github.io/dev_guide/building.html#crangotchas) helpful, and I'd be happy to provide support through that process.
