### 2019-03-23 ab0cab9
update tests & vignette
