+++ R/download_ramlegacy.R
@@ -1,5 +1,6 @@
 
 #' @name download_ramlegacy
+#' @family ram
 #' @title Download RAM Legacy Excel Database
 #' @description Downloads a specified version of RAM Legacy Stock Assessment
 #'  Excel Database and as an RDS object to a local directory specified by \code{\link{ram_dir}}.
@@ -17,10 +18,11 @@
 #'  package and can be viewed by calling \code{\link{ram_dir}}. Although this is not the \strong{recommended}
 #'  approach \code{download_ramlegacy} supports downloading to a user-specified path.
 #' @param ram_url A string. By default it is set to the zenodo url of the database.
-#'  Please \strong{do not pass} in any other url to
-#'  \code{ram_url}.
+#' Please \strong{do not pass} in any other url to
+#' \code{ram_url}.
 #' @param overwrite If TRUE, user will not encounter the usual interactive prompt confirming whether they want
-#'  overwrite the version present locally.
+#' overwrite the version present locally.
+#' @param quiet If TRUE, suppress status messages
 #' @export
 #' @examples
 #' \dontrun{
@@ -36,7 +38,10 @@
 #' download_ramlegacy(version = "4.3")
 #' }
 download_ramlegacy <- function(version = NULL, ram_path = NULL,
-ram_url = "https://doi.org/10.5281/zenodo.2542918", overwrite = FALSE) {
+ram_url = "https://doi.org/10.5281/zenodo.2542918", overwrite = FALSE, quiet = FALSE) {
+
+  # check internet connection and throw error if there is a connection issue
+  net_check(ram_url, show_error = TRUE)
 
   # user_path, a boolean flag set to FALSE by default
   user_path <- FALSE
@@ -91,12 +96,11 @@ ram_url = "https://doi.org/10.5281/zenodo.2542918", overwrite = FALSE) {
     dir.create(vers_path, recursive = TRUE)
   }
 
-  # check internet connection and throw error if there is a connection issue
-  net_check(ram_url, show_error = TRUE)
-
 
   # notify the user
-  notify("Downloading...this may take a while")
+  if(!quiet) {
+    notify("Downloading...this may take a while")
+  }
 
   # constructing different urls and downloading depending on whether old
   # or recent version
@@ -121,7 +125,9 @@ ram_url = "https://doi.org/10.5281/zenodo.2542918", overwrite = FALSE) {
     httr::GET(ram_url, httr::write_disk(excel_path))
 
     if (file.exists(excel_path)) {
-      notify("Finished downloading. Saving the database as RDS object...")
+      if (!quiet) {
+        notify("Finished downloading. Saving the database as RDS object...")
+      }
       suppressWarnings(read_ramlegacy(vers_path, version))
     }
 
@@ -140,12 +146,16 @@ ram_url = "https://doi.org/10.5281/zenodo.2542918", overwrite = FALSE) {
 
     # unzip it and read it in
     if (file.exists(tmp)) {
-      notify(paste(
-        "Finished Downloading.",
-        "Now unzipping it..."
-      ))
+      if (!quiet) {
+        notify(paste(
+          "Finished Downloading.",
+          "Now unzipping it..."
+        ))
+      }
       utils::unzip(tmp, exdir = vers_path, overwrite = TRUE)
-      notify("Saving the unzipped database as RDS object...")
+      if (!quiet) {
+        notify("Saving the unzipped database as RDS object...")
+      }
       vers_path <- file.path(vers_path, paste0("RLSADB v", version))
       vers_path <- file.path(vers_path, "DB Files With Assessment Data")
       suppressWarnings(read_ramlegacy(vers_path, version))
