+++ tests/testthat/test_load_ramlegacy.R
@@ -1,11 +1,3 @@
 context("test load_ramlegacy works as expected")
 
-# test_that("load_ramlegacy errors out when version is not present locally", {
-#   temp_path <- tempfile("ramlegacy", tempdir())
-#   # pass in the path to rds to load_ramlegacy
-#   rds_path <- file.path(temp_path, "1.0/v1.0.rds")
-#   expect_error(
-#     load_ramlegacy("1.0", rds_path),
-#     "Version 1.0 not found locally."
-#   )
-# })
+
