+++ tests/testthat/test_download_ramlegacy.R
@@ -1,184 +1,63 @@
 context("Testing that download_ramlegacy works")
 
-# test_that("defaults to curr. latest version 4.3 if version not specified", {
-#   skip_on_cran()
-#   temp_path <- tempfile("ramlegacy", tempdir())
-#   download_ramlegacy(NULL, temp_path)
-#   vers_path <- file.path(temp_path, "4.44")
-#   rds_path <- file.path(vers_path, "v4.3.rds")
-#   expect_true(file.exists(rds_path))
-#   unlink(rds_path, recursive = TRUE)
-# })
+test_that("defaults to curr. latest version 4.44 if version not specified", {
+  #skip_on_cran()
+  temp_path <- tempfile("ramlegacy", tempdir())
+  download_ramlegacy(NULL, temp_path)
+  vers_path <- file.path(temp_path, "4.44")
+  rds_path <- file.path(vers_path, "v4.44.rds")
+  expect_true(file.exists(rds_path))
+  unlink(rds_path, recursive = TRUE)
+})
 
-# test_that("download_ramlegacy errors when there is no internet", {
-#   temp_path <- tempfile("ramlegacy", tempdir())
-#   httptest::without_internet(
-#     expect_error(
-#       download_ramlegacy("3.0", temp_path),
-#       paste(
-#         "Could not connect to the internet.",
-#         "Please check your connection settings and try again."
-#       )
-#     )
-#   )
-#   unlink(temp_path, recursive = TRUE)
-# })
-#
-# test_that("download_ramlegacy errors out behind a proxy server", {
-#   skip_on_cran()
-#   temp_path <- tempfile("ramlegacy", tempdir())
-#   httr::with_config(
-#     httr::use_proxy(url = "http://google.com", port = 1234),
-#     expect_error(
-#       download_ramlegacy("3.0", temp_path),
-#       paste(
-#         "Could not connect to the internet.",
-#         "Please check your connection settings and try again."
-#       )
-#     )
-#   )
-#   unlink(temp_path, recursive = TRUE)
-# })
+test_that("download_ramlegacy errors when there is no internet", {
+  temp_path <- tempfile("ramlegacy", tempdir())
+  httptest::without_internet(
+    expect_error(
+      download_ramlegacy("3.0", temp_path),
+      paste(
+        "Could not connect to the internet.",
+        "Please check your connection settings and try again."
+      )
+    )
+  )
+  unlink(temp_path, recursive = TRUE)
+})
 
-# test_that("download_ramlegacy downloads from backup when website is down", {
-#   skip_on_cran()
-#   temp_path <- tempfile("ramlegacy", tempdir())
-#
-#   test_url1 <- "http://httpbin.org/status/300"
-#   test_url3 <- "http://httpbin.org/status/404"
-#   test_url4 <- "http://httpbin.org/status/500"
-#
-#   expect_message(
-#     download_ramlegacy("3.0", temp_path, test_url1),
-#     "Downloading from backup location..."
-#   )
-#   unlink(temp_path, recursive = TRUE)
-#   expect_message(
-#     download_ramlegacy("2.0", temp_path, test_url3),
-#     "Downloading from backup location..."
-#   )
-#   unlink(temp_path, recursive = TRUE)
-#   expect_message(
-#     download_ramlegacy("1.0", temp_path, test_url4),
-#     "Downloading from backup location..."
-#   )
-#   unlink(temp_path, recursive = TRUE)
-# })
+test_that("download_ramlegacy errors out behind a proxy server", {
[…]
