+++ man/load_ramlegacy.Rd
@@ -43,4 +43,8 @@ load_ramlegacy(version = "2.5")
 }
 
 }
+\seealso{
+Other ram: \code{\link{download_ramlegacy}},
+  \code{\link{ram_dir}}
+}
 \concept{ram}
