### 2019-03-23 c421712
Update load_ramlegacy.R
