+++ vignettes/ramlegacy.Rmd
@@ -9,8 +9,6 @@ vignette: |
   %\VignetteEncoding{UTF-8}
 ---
 
-
-
 ```{r setup, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
@@ -21,12 +19,11 @@ knitr::opts_chunk$set(
 From the database's [website](www.ramlegacy.org):
 
 > The RAM Legacy Stock Assessment Database is a compilation of stock assessment results for commercially exploited marine populations from around the world. It is inspired by Dr. Ransom A. Myers' original [stock-recruitment database](http://www.mathstat.dal.ca/~myers/welcome.html), which is no longer being updated."
-
-Different versions of the database are available in compressed (zipped) form on the [website](www.ramlegacy.org) in two formats: **Access** (Microsoft Office Access 2007 file format)  and **Excel** (Microsoft Excel Open XML Format Spreadsheet). As of January 2019, the available versions are 1.0, 2.0, 2.5, 3.0, 4.3, and 4.4.
+Different versions of the database are available in compressed (zipped) form on the [website](www.ramlegacy.org) in two formats: **Access** (Microsoft Office Access 2007 file format)  and **Excel** (Microsoft Excel Open XML Format Spreadsheet). As of March 2019, the available versions are 1.0, 2.0, 2.5, 3.0, 4.3, and 4.40, 4.41, 4.44.
 
 ## About the package
 The goal of `ramlegacy` is to cache the specified version/versions of the RAM Legacy Stock Assessment Excel Database and this way save time and effort spent in re-downloading different
-versions of the database as part of any future analysis involving the data. The package also supports reading in any specified cached version of the database into the user's environment through `library(ramlegacy)` and a dedicated function called `load_ramlegacy()`. The user can always view the location to which the database is cached and read from by running `ram_dir()` and specifying its version argument. This vignette will outline a few more details about these functions and the package that will hopefully make `ramlegacy` easy to use.
+versions of the database as part of any future analysis involving the data. The package also supports reading in any specified dataframe(s) of the database through a function called `load_ramlegacy()`. The user can always view the location to which the database is cached and read from by running `ram_dir()` and specifying its version argument. This vignette will outline a few more details about these functions and the package that will hopefully make `ramlegacy` easy to use.
 
 ## Package Installation
 You can install the development version from [GitHub](https://github.com/kshtzgupta1/ramlegacy) with the `install_github` function available in `devtools` package:
@@ -39,39 +36,18 @@ install_github("kshtzgupta1/ramlegacy")
 
 To ensure that the vignette is installed along with the package make sure to remove `--no-build-vignettes` from the `build_opts` in `install_github`
 
-## Package loading
-
-```{r packages, warning=FALSE, message=FALSE, echo = TRUE, eval = F}
-library(ramlegacy)
-```
-
-After installation of the package when `library(ramlegacy)` is called for the first time `ramlegacy` checks whether any version of the database exists in the user's application data directory whose path is given by `ram_dir()`. If no version of the database is found then the user is informed about it and prompted to download a version by calling `download_ramlegacy()`.
-
-Subsequent calls to `library(ramlegacy)` read in and load the database in the user's global environment. Which version is loaded depends upon which versions have been downloaded and are present locally and whether `library(ramlegacy)` is called in an interactive session or non-interactive session.
[…]
