### 2019-03-23 cb995ec
Update test_read_ramlegacy.R
