+++ vignettes/ramlegacy.Rmd
@@ -69,7 +69,7 @@ To enable caching behavior, before downloading a the specified version `download
 Use `load_ramlegacy` to obtain a list of particular dataframes or of all the dataframes present in the desired version of the database. Make sure that the version you want to load was downloaded and is already present locally otherwise `load_ramlegacy` will throw an error.
 
 ```{r, load_ramlegacy_example1, echo = T, eval = F}
-# loads in area and bioparams tables from version 4.44 database
+# returns a list containing area and bioparams tables from version 4.44 database
 load_ramlegacy(version = "4.44", tables = c("area", "bioparams"))
 ```
 
