### 2019-03-26 1b066c8
update download tests
