+++ tests/testthat/test_download_ramlegacy.R
@@ -5,7 +5,9 @@ test_that("defaults to curr. latest version 4.44 if version not specified", {
   temp_path <- tempfile("ramlegacy", tempdir())
   download_ramlegacy(NULL, temp_path)
   vers_path <- file.path(temp_path, "4.44")
-  rds_path <- file.path(vers_path, "v4.44.rds")
+  rds_path <- file.path(vers_path, "RLSADB v4.44")
+  rds_path <- file.path(rds_path, "DB Files With Assessment Data")
+  rds_path <- file.path(rds_path, "v4.44.rds")
   expect_true(file.exists(rds_path))
   unlink(rds_path, recursive = TRUE)
 })
@@ -25,7 +27,7 @@ test_that("download_ramlegacy errors when there is no internet", {
 })
 
 test_that("download_ramlegacy errors out behind a proxy server", {
-  skip_on_cran()
+  #skip_on_cran()
   temp_path <- tempfile("ramlegacy", tempdir())
   httr::with_config(
     httr::use_proxy(url = "http://google.com", port = 1234),
@@ -40,20 +42,18 @@ test_that("download_ramlegacy errors out behind a proxy server", {
   unlink(temp_path, recursive = TRUE)
 })
 
-
-test_that("Doesn't download when requested version is already present", {
-  skip_on_cran()
+test_that("Overwrites existing download when overwrite = TRUE", {
   temp_path <- tempfile("ramlegacy", tempdir())
-  # download version 3.0 for the first time
-  download_ramlegacy("3.0", temp_path)
+  # download version 4.44 for the first time
+  download_ramlegacy("4.44", temp_path)
   # call download_ramlegacy again to test behavior
-  expect_is(download_ramlegacy("3.0", temp_path), "character")
+  expect_true(download_ramlegacy("4.44", temp_path, overwrite = TRUE))
   unlink(temp_path, recursive = TRUE)
 })
 
 # testing download_ramlegacy downloads older version 1.0
 test_that("download_ramlegacy downloads v1.0", {
-  skip_on_cran()
+  #skip_on_cran()
   temp_path <- tempfile("ramlegacy", tempdir())
   download_ramlegacy("1.0", temp_path)
   vers_path <- file.path(temp_path, "1.0")
