+++ tests/testthat/test_read_ramlegacy.R
@@ -3,7 +3,7 @@ context("Checking read_ramlegacy works as expected")
 read_url <- "https://zenodo.org/record/2542919/files/RLSADB%20v4.44.zip?download=1"
 
 test_that("read_ramlegacy writes the database as RDS for v4.44", {
-  #skip_on_cran()
+  skip_on_cran()
   temp_db_path <- file.path(tempfile("ramlegacy", tempdir()), "4.44")
   dir.create(temp_db_path, showWarnings = FALSE, recursive = TRUE)
   # download and unzip the database
