+++ R/load_ramlegacy.R
@@ -24,9 +24,8 @@
 #'
 #' @examples
 #' \donttest{
-#' # If version is not specified then current latest version
-#' # returns a list containing area and bioparams tables from version 4.44 database
-# load_ramlegacy(version = "4.44", tables = c("area", "bioparams"))
+#' # Returns a list containing area and bioparams tables from version 4.44 database
+#' load_ramlegacy(version = "4.44", tables = c("area", "bioparams"))
 #' }
 load_ramlegacy <- function(version = NULL, tables = NULL, ram_path = NULL) {
   ram_url <- "https://doi.org/10.5281/zenodo.2542918"
