### 2019-04-03 ed955c9
fix ramlist print
