+++ tests/testthat/test_load_ramlegacy.R
@@ -14,7 +14,7 @@ test_that("load_ramlegacy works as expected when particular dataframes are reque
   temp_path <- tempfile("ramlegacy", tempdir())
   download_ramlegacy("4.44", temp_path)
   vers_path <- file.path(temp_path, "4.44")
-  dir_path <- file.path(vers_path, paste0("RLSADB v", version))
+  dir_path <- file.path(vers_path, "RLSADB v4.44")
   subdir_path <- file.path(dir_path, "DB Files With Assessment Data")
   rds_path <- file.path(subdir_path, "v4.44.rds")
   lst_dfs <- load_ramlegacy("4.44", tables = c("area", "bioparams"), ram_path = rds_path)
@@ -26,7 +26,8 @@ test_that("load_ramlegacy works as expected when all dataframes are requested fo
   temp_path <- tempfile("ramlegacy", tempdir())
   download_ramlegacy("4.44", temp_path)
   vers_path <- file.path(temp_path, "4.44")
-  subdir_path <- file.path(vers_path, "DB Files With Assessment Data")
+  dir_path <- file.path(vers_path, "RLSADB v4.44")
+  subdir_path <- file.path(dir_path, "DB Files With Assessment Data")
   rds_path <- file.path(subdir_path, "v4.44.rds")
   lst_dfs <- load_ramlegacy("4.44", ram_path = rds_path)
   expect_equal(length(lst_dfs), 26)
