### 2019-04-03 d730af3
fix final style issues
