### 2019-04-03 1e77fad
add ropensci footer
