+++ README.Rmd
@@ -120,3 +120,5 @@ Use of the RAM Legacy Stock Assessment Database is subject to a [Fair Use Policy
 Please cite the RAM Legacy Stock Assessment Database as follows:
 
 Ricard, D., Minto, C., Jensen, O.P. and Baum, J.K. (2012) Evaluating the knowledge base and status of commercially exploited marine species with the RAM Legacy Stock Assessment Database. Fish and Fisheries 13 (4) 380-398. DOI: 10.1111/j.1467-2979.2011.00435.x
+
+[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
