### 2019-04-03 #264 @kshtzgupta1
@stefaniebutland Thank you so much for reaching out. My sincere apologies for the late reply. I think Prof. Boettiger and I would definitely be interested in authoring a blog post about ramlegacy.  I'll discuss the content of the post with Prof. Boettiger and let you know about a timeline very soon.
