### 2019-04-03 #264 @jafflerbach
Yep both of those are correct.

Jamie Afflerbach
Marine Data Scientist
National Center for Ecological Analysis and Synthesis (NCEAS
<https://www.nceas.ucsb.edu/>)
University of California, Santa Barbara
website <http://jamieafflerbach.com> - github
<https://github.com/jafflerbach>- twitter <https://twitter.com/jafflerbach>

On Wed, Apr 3, 2019 at 1:39 PM Kshitiz Gupta <notifications@github.com>
wrote:

> @jafflerbach <https://github.com/jafflerbach> Thank you very much for
> reviewing the package! I would like to acknowledge you as a "rev"-type
> contributor in the Authors@R field if that's okay with you. Can I put in
> your email as afflerbach@nceas.ucsb.edu and orcid id as
> https://orcid.org/0000-0002-5215-9342 ?
>
