+++ man/ramlegacy-package.Rd
@@ -23,9 +23,9 @@ Fish and Fisheries 13 (4) 380-398. <doi:10.1111/j.1467-2979.2011.00435.x.>
 \seealso{
 \url{www.ramlegacy.org}
 
-\url{www.https://github.com/kshtzgupta1/ramlegacy}
+\url{www.https://github.com/ropensci/ramlegacy}
 
-\url{www.https://github.com/kshtzgupta1/ramlegacy/issues}
+\url{www.https://github.com/ropensci/ramlegacy/issues}
 }
 \author{
 \strong{Maintainer}: Kshitiz Gupta \email{kshtzgupta1@berkeley.edu} [copyright holder]
