+++ _pkgdown.yml
@@ -20,7 +20,7 @@ navbar:
       href: articles/ramlegacy.html
     github:
       icon: fa-github fa-lg
-      href: https://github.com/kshtzgupta1/ramlegacy
+      href: https://github.com/ropensci/ramlegacy
 
 reference:
 - title: All functions
