+++ R/ramlegacy-package.R
@@ -1,10 +1,17 @@
+#' ramlegacy: download, cache and read RAM Legacy Stock Assessment Database
+#'
+#' ramlegacy contains functions to download, cache and read in specified tables
+#' from the excel version of the RAM Legacy Stock Assessment Database,
+#' an online compilation of stock assessment results for commercially exploited
+#' marine populations from around the world. More information about the database
+#' can be found at <http://ramlegacy.org/>.
 #' @keywords internal
 #' @references Ricard, D., Minto, C., Jensen, O.P. and Baum, J.K. (2012)
 #' Evaluating the knowledge base and status of commercially exploited marine
 #' species with the RAM Legacy Stock Assessment Database.
 #' Fish and Fisheries 13 (4) 380-398. <doi:10.1111/j.1467-2979.2011.00435.x.>
 #' @seealso \url{www.ramlegacy.org}
-#' @seealso \url{www.https://github.com/ropensci/ramlegacy}
-#' @seealso \url{www.https://github.com/ropensci/ramlegacy/issues}
+#' @seealso \url{www.github.com/ropensci/ramlegacy}
+#' @seealso \url{www.github.com/ropensci/ramlegacy/issues}
 
 "_PACKAGE"
