+++ README.Rmd
@@ -21,8 +21,8 @@ ramlegacy
 
 - **Authors**: Kshitiz Gupta, [Carl Boettiger](http://www.carlboettiger.info/)
 - **License**: [MIT](http://opensource.org/licenses/MIT)
-- [Package source code on Github](https://github.com/kshtzgupta1/ramlegacy)
-- [**Submit Bugs and feature requests**](https://github.com/kshtzgupta1/ramlegacy/issues)
+- [Package source code on Github](https://github.com/ramlegacy/ramlegacy)
+- [**Submit Bugs and feature requests**](https://github.com/ramlegacy/ramlegacy/issues)
 
 `ramlegacy` is an R package that supports caching and reading in different versions of the RAM Legacy Stock Assessment Data Base, an online compilation of stock assessment results for commercially exploited marine populations from around the world. More information about the database can be found [here.](http://ramlegacy.org)
 
@@ -36,18 +36,18 @@ ramlegacy
   - Provides a function `ram_dir()` to view the path where the downloaded database was saved.
 
 ## Installation
-You can install the development version from [GitHub](https://github.com/kshtzgupta1/ramlegacy) with:
+You can install the development version from [GitHub](https://github.com/ramlegacy/ramlegacy) with:
 
 ```{r, eval = FALSE, echo = TRUE}
 install.packages("devtools")
 library(devtools)
-install_github("kshtzgupta1/ramlegacy")
+install_github("ramlegacy/ramlegacy")
 ```
 
 To ensure that the vignette is installed along with the package make sure to remove `--no-build-vignettes` from the `build_opts` in `install_github`
 
 ## Usage
-Please see the ramlegacy [vignette](https://kshtzgupta1.github.io/ramlegacy/articles/ramlegacy.html) for more detailed examples and additional package functionality. 
+Please see the ramlegacy [vignette](https://ramlegacy.github.io/ramlegacy/articles/ramlegacy.html) for more detailed examples and additional package functionality. 
 
 Start by loading the package using `library`.
 
@@ -55,7 +55,7 @@ Start by loading the package using `library`.
 library(ramlegacy)
 ```
 
-When `ramlegacy` is loaded for the first time after installation of the package calling `library(ramlegacy)` will prompt the user to download a version of the database using `download_ramlegacy()`. After downloading a version or multiple versions of the database the subsequent behavior of `library(ramlegacy)` will depend on which version/versions were downloaded and are present on disk as well as whether `library(ramlegacy)` is called in an interactive vs non-interactive session. For more details about this behavior please see the ramlegacy [vignette.](https://kshtzgupta1.github.io/ramlegacy/articles/ramlegacy.html) 
+When `ramlegacy` is loaded for the first time after installation of the package calling `library(ramlegacy)` will prompt the user to download a version of the database using `download_ramlegacy()`. After downloading a version or multiple versions of the database the subsequent behavior of `library(ramlegacy)` will depend on which version/versions were downloaded and are present on disk as well as whether `library(ramlegacy)` is called in an interactive vs non-interactive session. For more details about this behavior please see the ramlegacy [vignette.](https://ramlegacy.github.io/ramlegacy/articles/ramlegacy.html) 
 
 ### download_ramlegacy
 
@@ -94,11 +94,11 @@ load_ramlegacy()
 ### ram_dir
 To view the exact path where a certain version of the database was downloaded and cached by `download_ramlegacy` you can run `ram_dir(vers = 'version')`, specifying the version number inside the function call:
[…]
