+++ README.Rmd
@@ -24,30 +24,31 @@ ramlegacy
 - [Package source code on Github](https://github.com/ramlegacy/ramlegacy)
 - [**Submit Bugs and feature requests**](https://github.com/ramlegacy/ramlegacy/issues)
 
-`ramlegacy` is an R package that supports caching and reading in different versions of the RAM Legacy Stock Assessment Data Base, an online compilation of stock assessment results for commercially exploited marine populations from around the world. More information about the database can be found [here.](http://ramlegacy.org)
+`ramlegacy` is an R package that supports caching and reading in different versions of the RAM Legacy Stock Assessment Data Base, an online compilation of stock assessment results for commercially exploited marine populations from around the world. More information about the database can be found [here.](https://www.ramlegacy.org/)
 
 
 ## What does `ramlegacy` do?
 
   - Provides a function `download_ramlegacy()`, to download all the available  
-    versions of the RAM Legacy Stock Assessment Excel Database as RDS objects. This way once
-    a version has been downloaded it doesn't need to be re-downloaded for subsequent analysis.
-  - Supports reading in the cached versions of the database through loading the package i.e.         calling `library(ramlegacy)` and also by providing a function `load_ramlegacy()` to load any       specified version.
-  - Provides a function `ram_dir()` to view the path where the downloaded database was saved.
+    versions of the RAM Legacy Stock Assessment Excel Database and cache them on the user's computer as           serialized RDS objects. This way once a version has been downloaded it doesn't need to be re-downloaded
+    for subsequent analysis.
+  - Supports reading in specified tables or all tables from a cached version of the database through a 
+    function `load_ramlegacy()`
+  - Provides a function `ram_dir()` to view the path of the location where the downloaded database was cached.
 
 ## Installation
-You can install the development version from [GitHub](https://github.com/ramlegacy/ramlegacy) with:
+You can install the development version from [gitHub](https://github.com/ropensci/ramlegacy) with:
 
 ```{r, eval = FALSE, echo = TRUE}
 install.packages("devtools")
 library(devtools)
-install_github("ramlegacy/ramlegacy")
+install_github("ropensci/ramlegacy")
 ```
 
 To ensure that the vignette is installed along with the package make sure to remove `--no-build-vignettes` from the `build_opts` in `install_github`
 
 ## Usage
-Please see the ramlegacy [vignette](https://ramlegacy.github.io/ramlegacy/articles/ramlegacy.html) for more detailed examples and additional package functionality. 
+Please see the ramlegacy [vignette](https://ropensci.github.io/ramlegacy/articles/ramlegacy.html) for more detailed examples and additional package functionality. 
 
 Start by loading the package using `library`.
 
@@ -55,39 +56,48 @@ Start by loading the package using `library`.
 library(ramlegacy)
 ```
 
-When `ramlegacy` is loaded for the first time after installation of the package calling `library(ramlegacy)` will prompt the user to download a version of the database using `download_ramlegacy()`. After downloading a version or multiple versions of the database the subsequent behavior of `library(ramlegacy)` will depend on which version/versions were downloaded and are present on disk as well as whether `library(ramlegacy)` is called in an interactive vs non-interactive session. For more details about this behavior please see the ramlegacy [vignette.](https://ramlegacy.github.io/ramlegacy/articles/ramlegacy.html) 
-
 ### download_ramlegacy
 
[…]
