+++ README.md
@@ -7,28 +7,28 @@ ramlegacy
 
 -   **Authors**: Kshitiz Gupta, [Carl Boettiger](http://www.carlboettiger.info/)
 -   **License**: [MIT](http://opensource.org/licenses/MIT)
--   [Package source code on Github](https://github.com/kshtzgupta1/ramlegacy)
--   [**Submit Bugs and feature requests**](https://github.com/kshtzgupta1/ramlegacy/issues)
+-   [Package source code on Github](https://github.com/ramlegacy/ramlegacy)
+-   [**Submit Bugs and feature requests**](https://github.com/ramlegacy/ramlegacy/issues)
 
-`ramlegacy` is an R package that supports caching and reading in different versions of the RAM Legacy Stock Assessment Data Base, an online compilation of stock assessment results for commercially exploited marine populations from around the world. More information about the database can be found [here.](http://ramlegacy.org)
+`ramlegacy` is an R package that supports caching and reading in different versions of the RAM Legacy Stock Assessment Data Base, an online compilation of stock assessment results for commercially exploited marine populations from around the world. More information about the database can be found [here.](https://www.ramlegacy.org/)
 
 What does `ramlegacy` do?
 -------------------------
 
 -   Provides a function `download_ramlegacy()`, to download all the available
-    versions of the RAM Legacy Stock Assessment Excel Database as RDS objects. This way once a version has been downloaded it doesn't need to be re-downloaded for subsequent analysis.
--   Supports reading in the cached versions of the database through loading the package i.e. calling `library(ramlegacy)` and also by providing a function `load_ramlegacy()` to load any specified version.
--   Provides a function `ram_dir()` to view the path where the downloaded database was saved.
+    versions of the RAM Legacy Stock Assessment Excel Database and cache them on the user's computer as serialized RDS objects. This way once a version has been downloaded it doesn't need to be re-downloaded for subsequent analysis.
+-   Supports reading in specified tables or all tables from a cached version of the database through a function `load_ramlegacy()`
+-   Provides a function `ram_dir()` to view the path of the location where the downloaded database was cached.
 
 Installation
 ------------
 
-You can install the development version from [GitHub](https://github.com/kshtzgupta1/ramlegacy) with:
+You can install the development version from [gitHub](https://github.com/ropensci/ramlegacy) with:
 
 ``` r
 install.packages("devtools")
 library(devtools)
-install_github("kshtzgupta1/ramlegacy")
+install_github("ropensci/ramlegacy")
 ```
 
 To ensure that the vignette is installed along with the package make sure to remove `--no-build-vignettes` from the `build_opts` in `install_github`
@@ -36,7 +36,7 @@ To ensure that the vignette is installed along with the package make sure to rem
 Usage
 -----
 
-Please see the ramlegacy [vignette](https://kshtzgupta1.github.io/ramlegacy/articles/ramlegacy.html) for more detailed examples and additional package functionality.
+Please see the ramlegacy [vignette](https://ropensci.github.io/ramlegacy/articles/ramlegacy.html) for more detailed examples and additional package functionality.
 
 Start by loading the package using `library`.
 
@@ -44,40 +44,49 @@ Start by loading the package using `library`.
 library(ramlegacy)
 ```
 
[…]
