+++ R/load_ramlegacy.R
@@ -1,30 +1,114 @@
 #' @name load_ramlegacy
 #' @family ramlegacy functions
-#' @title Read-in downloaded RAM Legacy Database
-#' @description Returns a list of specific dataframes or a list of all the dataframes present in the
-#'  requested version of the database.
-#' @param version A character vector of length 1 specifying the version number of the
-#'  database. As of February 2019, the available versions are "1.0", "2.0", "2.5", "3.0", "4.3",
-#'  "4.40", "4.41" and "4.44". If version argument is not specified then it defaults to latest version
-#'  (currently "4.44"). If you want to load multiple versions please
-#'  load them one at a time as passing them all at once will throw an error.
-#' @param tables A character vector specifying the names of particular dataframes to load from a
-#'  certain version. If not specified then a list containing all the dataframes within the requested version
-#'  is returned.
+#' @title Read-in dataframes from downloaded RAM Legacy Database
+#' @description Returns a list of specific dataframes or a list of all the
+#'  dataframes present in the requested version of the database.
+#' @param version A character vector of length 1 specifying the version number
+#' of the database. As of writing this package, the available versions are
+#' "1.0", "2.0", "2.5", "3.0", "4.3", "4.40", "4.41" and "4.44". If version
+#' argument is not specified then it defaults to newest version (v4.44).
+#' @param tables A character vector specifying the names of particular
+#' dataframes to load from a specified version. If not specified then
+#' a list containing all the dataframes within the requested version
+#' is returned. For a description of the different tables present
+#' in the database please see below.
 #' @param ram_path path to the local directory where the specified version of
-#' the RAM Legacy Stock Excel Assessment Database was downloaded.
-#' By default this path is set to within the rapddirs directory and can be viewed using calling the function
-#' \code{\link{ram_dir}} and specifying the version number inside.
-#' Although this is not the \strong{recommended} approach \code{load_ramlegacy} supports reading in the
-#'  database's dataframes from a user-specified path provided that the database is present
-#'  at the specified path as an rds object.
+#' the RAM Legacy Stock Excel Assessment Database was downloaded. By default
+#' this path is set to within the rapddirs directory and can be viewed using
+#' calling the function \code{\link{ram_dir}} and specifying the version number
+#' inside the function call. Although this is not the \strong{recommended}
+#' approach \code{load_ramlegacy} supports reading in the database's
+#' dataframes from a user-specified path provided that the database is present
+#' at the specified path as an rds object.
 #' @export
 #'
 #' @examples
 #' \donttest{
-#' # Returns a list containing area and bioparams tables from version 4.44 database
+#' #' Returns a list containing area and bioparams tables from version 4.44 database
 #' load_ramlegacy(version = "4.44", tables = c("area", "bioparams"))
 #' }
+#'
+#' @section Description of the dataframes present in the database:
+#' \itemize{
+#' \item metadata: Table with summarized metadata (only available in newer
+#'  versions 4.40, 4.41, 4.44)
+#' \item stock: This stores the stock database table
+#' \item assessment: This stores the assessment database table
+#' \item taxonomy: This stores the taxonomy database table
+#' \item management: This stores the management database table
+#' \item assessor: This stores the assessor database table
+#' \item assessmetod: This stores the assessmetod database table
+#' \item area: This stores the area database table
[…]
