+++ man/load_ramlegacy.Rd
@@ -2,38 +2,90 @@
 % Please edit documentation in R/load_ramlegacy.R
 \name{load_ramlegacy}
 \alias{load_ramlegacy}
-\title{Read-in downloaded RAM Legacy Database}
+\title{Read-in dataframes from downloaded RAM Legacy Database}
 \usage{
 load_ramlegacy(version = NULL, tables = NULL, ram_path = NULL)
 }
 \arguments{
-\item{version}{A character vector of length 1 specifying the version number of the
-database. As of February 2019, the available versions are "1.0", "2.0", "2.5", "3.0", "4.3",
-"4.40", "4.41" and "4.44". If version argument is not specified then it defaults to latest version
-(currently "4.44"). If you want to load multiple versions please
-load them one at a time as passing them all at once will throw an error.}
+\item{version}{A character vector of length 1 specifying the version number
+of the database. As of writing this package, the available versions are
+"1.0", "2.0", "2.5", "3.0", "4.3", "4.40", "4.41" and "4.44". If version
+argument is not specified then it defaults to newest version (v4.44).}
 
-\item{tables}{A character vector specifying the names of particular dataframes to load from a
-certain version. If not specified then a list containing all the dataframes within the requested version
-is returned.}
+\item{tables}{A character vector specifying the names of particular
+dataframes to load from a specified version. If not specified then
+a list containing all the dataframes within the requested version
+is returned. For a description of the different tables present
+in the database please see below.}
 
 \item{ram_path}{path to the local directory where the specified version of
-the RAM Legacy Stock Excel Assessment Database was downloaded.
-By default this path is set to within the rapddirs directory and can be viewed using calling the function
-\code{\link{ram_dir}} and specifying the version number inside.
-Although this is not the \strong{recommended} approach \code{load_ramlegacy} supports reading in the
- database's dataframes from a user-specified path provided that the database is present
- at the specified path as an rds object.}
+the RAM Legacy Stock Excel Assessment Database was downloaded. By default
+this path is set to within the rapddirs directory and can be viewed using
+calling the function \code{\link{ram_dir}} and specifying the version number
+inside the function call. Although this is not the \strong{recommended}
+approach \code{load_ramlegacy} supports reading in the database's
+dataframes from a user-specified path provided that the database is present
+at the specified path as an rds object.}
 }
 \description{
-Returns a list of specific dataframes or a list of all the dataframes present in the
- requested version of the database.
+Returns a list of specific dataframes or a list of all the
+ dataframes present in the requested version of the database.
 }
+\section{Description of the dataframes present in the database}{
+
+\itemize{
+\item metadata: Table with summarized metadata (only available in newer
+ versions 4.40, 4.41, 4.44)
+\item stock: This stores the stock database table
+\item assessment: This stores the assessment database table
+\item taxonomy: This stores the taxonomy database table
+\item management: This stores the management database table
+\item assessor: This stores the assessor database table
+\item assessmetod: This stores the assessmetod database table
+\item area: This stores the area database table
+\item biometrics: This stores the biometrics database table
+\item tsmetrics: This stores the tsmetrics database table
+\item timeseries: The time series data is a matrix object with the following
+ headers/columns: (1) assessid (2) stockid (3) stocklong (4) tsid (5) tsyear
+ (6) tsvalue
+\item bioparams: The parameter data is a matrix object with the following
+headers/columns: (1) assessid (2) stockid (3) stocklong (4) bioid (5) biovalue
+(6) bioyear (7) bionotes
+\item timeseries_values_views: This stores the timeseries values with timeseries
+type along the columns and stocks along the rows
+\item timeseries_units_views: This stores the timeseries values with timeseries
+type along the columns and stocks along the rows
[…]
