+++ man/load_ramlegacy.Rd
@@ -33,6 +33,7 @@ Returns a list of specific dataframes or a list of all the
 }
 \section{Description of the dataframes present in the database}{
 
+
 \itemize{
 \item metadata: Table with summarized metadata (only available in newer
  versions 4.40, 4.41, 4.44)
@@ -46,8 +47,8 @@ Returns a list of specific dataframes or a list of all the
 \item biometrics: This stores the biometrics database table
 \item tsmetrics: This stores the tsmetrics database table
 \item timeseries: The time series data is a matrix object with the following
- headers/columns: (1) assessid (2) stockid (3) stocklong (4) tsid (5) tsyear
- (6) tsvalue
+headers/columns: (1) assessid (2) stockid (3) stocklong (4) tsid (5) tsyear
+(6) tsvalue
 \item bioparams: The parameter data is a matrix object with the following
 headers/columns: (1) assessid (2) stockid (3) stocklong (4) bioid (5) biovalue
 (6) bioyear (7) bionotes
@@ -63,7 +64,6 @@ timeseries type along the columns and stocks along the rows
 type along the columns and stocks along the rows
 \item timeseries_sources_views: This stores the timeseries sources with timeseries
 type along the columns and stocks along the rows
-}
 \item timeseries_years_views: This stores the timeseries years with timeseries
 type along the columns and stocks along the rows
 \item bioparams_values_views: This stores the reference point values, with
@@ -79,10 +79,47 @@ reference point type along the columns and stocks along the rows
 \item bioparams_notes_views: This stores the reference point notes, with
 reference point type along the columns and stocks along the rows
 }
+}
+
+\section{v4.40, v4.41, and v4.44 also contain tables for individual most-used time series}{
+
+\itemize{
+\item tb.data: Total Biomass
+\item ssb.data: Spawning Stock Biomass
+\item tn.data: Total Abundance
+\item r.data: Recruits
+\item tc.data: Total Catch
+\item tl.data: Total Landings
+\item recc.data: Recreational Catch
+\item f.data: Fishing Mortality
+\item er.data: Exploitation Rate
+\item divtb.data: TB/TBmsy
+\item divssb.data: SSB/SSBmsy
+\item ivf.data: F/Fmsy
+\item diver.data: ER/ERmsy
+\item divbpref.data: B/Bmsypref
+\item divupref.data: U/Umsypref
+\item tbbest.data: TBbest
+\item tcbest.data: TCbest
+\item erbest.data: ERbest
+\item divtb.mgt.data: TB/TBmgt
+\item divssb.mgt.data: SSB/SSBmgt
+\item divf.mgt.data: F/Fmgt
+\item diver.mgt.data: ER/ERmgt
+\item divbpref.mgt.data: B/Bmgtpref
+\item divupref.mgt.data: U/Umgtpref
+\item cpair.data: Cpair
+\item tac.data: TAC
+\item cadv.data: Cadvised
+\item survb.data: survB
+\item cpue.data: CPUE
+\item effort.data: EFFORT
+}
+}
 
 \examples{
 \donttest{
-#' Returns a list containing area and bioparams tables from version 4.44 database
+# Returns a list containing area and bioparams tables from version 4.44 database
 load_ramlegacy(version = "4.44", tables = c("area", "bioparams"))
 }
 
