### 2019-04-10 42a7270
update download and ram_dir doc
