+++ _pkgdown.yml
@@ -1,3 +1,10 @@
+title: ramlegacy
+url: https://ropensci.github.io/ramlegacy/
+template:
+  params:
+    bootswatch: flatly
+    ganalytics: UA-12345678-0
+
 navbar:
   structure:
     left:
