+++ vignettes/ramlegacy.Rmd
@@ -15,41 +15,53 @@ knitr::opts_chunk$set(
   comment = "#>"
 )
 ```
+
 ## RAM Legacy Stock Assessment Database
 From the database's [website](www.ramlegacy.org):
 
-> The RAM Legacy Stock Assessment Database is a compilation of stock assessment results for commercially exploited marine populations from around the world. It is inspired by Dr. Ransom A. Myers' original [stock-recruitment database](http://www.mathstat.dal.ca/~myers/welcome.html), which is no longer being updated."
-Different versions of the database are available in compressed (zipped) form on the [website](www.ramlegacy.org) in two formats: **Access** (Microsoft Office Access 2007 file format)  and **Excel** (Microsoft Excel Open XML Format Spreadsheet). As of March 2019, the available versions are 1.0, 2.0, 2.5, 3.0, 4.3, and 4.40, 4.41, 4.44.
+> The RAM Legacy Stock Assessment Database is a compilation of stock assessment results for commercially exploited marine populations from around the world. It is inspired by Dr. Ransom A. Myers' original [stock-recruitment database](http://wayback.archive-it.org/org-990/20170328023224/http://www.mscs.dal.ca/~myers/welcome.html), which is no longer being updated."
+
+As of April 2019, the available versions are 1.0, 2.0, 2.5, 3.0, 4.3, 4.40, 4.41, 4.44.
+
+The latest versions of the database (v4.44, v4.41, v4.40) are available at [Zenodo.](https://zenodo.org/communities/rlsadb/). This is where `download_ramlegacy` downloads the database from.
+
+The older versions of the database (v4.3, v3.0, v2.5, v2.0, v1.0) are available in **Excel** (Microsoft Excel Open XML Format Spreadsheet) format at this [github repo](https://github.com/kshtzgupta1/ramlegacy-assets).
 
 ## About the package
-The goal of `ramlegacy` is to cache the specified version/versions of the RAM Legacy Stock Assessment Excel Database and this way save time and effort spent in re-downloading different
-versions of the database as part of any future analysis involving the data. The package also supports reading in any specified dataframe(s) of the database through a function called `load_ramlegacy()`. The user can always view the location to which the database is cached and read from by running `ram_dir()` and specifying its version argument. This vignette will outline a few more details about these functions and the package that will hopefully make `ramlegacy` easy to use.
+The goal of `ramlegacy` is to cache a specified version(s) of the RAM Legacy Stock Assessment Excel Database and this way save time and effort spent in re-downloading different versions of the database as part of any future analysis involving the data. The package also supports reading in any specified dataframe(s) of the database through a function called `load_ramlegacy()`. The user can always view the location to which the database is cached and read from by running `ram_dir()` and specifying its version argument. This vignette will outline a few more details about these functions and the package that will hopefully make `ramlegacy` easy to use.
 
 ## Package Installation
-You can install the development version from [GitHub](https://github.com/kshtzgupta1/ramlegacy) with the `install_github` function available in `devtools` package:
+You can install the development version from [GitHub](https://github.com/ropensci/ramlegacy) with the `install_github` function available in `devtools` package:
 
 ```{r, eval = FALSE, echo = TRUE}
 install.packages("devtools")
 library(devtools)
-install_github("kshtzgupta1/ramlegacy")
+install_github("ropensci/ramlegacy")
 ```
 
 To ensure that the vignette is installed along with the package make sure to remove `--no-build-vignettes` from the `build_opts` in `install_github`
 
 ## download_ramlegacy
[…]
