+++ vignettes/tables_description.Rmd
@@ -105,7 +105,7 @@ knitr::opts_chunk$set(
     + TC, TL, RecC – Plots most recent catch and/or landings as well as recreational catch  and MSY if they         are also available from that assessment. (The year range of moratoriums are shaded grey if available.)
     + TAC, Cpair, Cadv – Plots most recent total allowable catch or landings (or other measures of total            quota), the corresponding catch or landings, and the corresponding scientifically advised catch if            available from the same assessment.
     + survB – Plots the most recent fishery-independent survey biomass time series.
-    + CPUE – Plots the most recent fishery-independend catch-per-unit-effort time series.
+    + CPUE – Plots the most recent fishery-independent catch-per-unit-effort time series.
     + EFFORT – Plots the most recent fishing effort time series.
     + CdivMSY – Plots the ratio of the most recent catch time series to MSY. MSY is drawn from the assessment       directly if available, or else is calculated from Bmsy and Umsy reference points (including proxies) if       available.
 
