### 2019-04-11 78678cd
fix doc prob in utils
