+++ man/ram_dir.Rd
@@ -21,12 +21,11 @@ in an OS independent manner. This is also the location from where
 }
 \examples{
 # return the path to the rappdirs directory where all version
-subdirectories are stored
+# subdirectories are stored
 ram_dir()
 
-# Returns the path of the subdirectory where version 4.3 of the database
-is downloaded to and
-# read from.
+# Returns the path of the subdirectory where version 4.3 of
+# the database is downloaded to and read from.
 ram_dir(vers = "4.3")
 }
 \seealso{
