+++ R/load_ramlegacy.R
@@ -24,10 +24,16 @@
 #'
 #' @examples
 #' \donttest{
-#' # Returns a list containing area and bioparams tables from version 4.44 database
+#' \dontshow{Sys.setenv(RAM_HOME = tempdir())}
+#'
+#' # first download version 4.44 of the database
+#' download_ramlegacy(version = "4.44")
+#'
+#' # get a list containing area and bioparams tables
+#' # from version 4.44 database
 #' load_ramlegacy(version = "4.44", tables = c("area", "bioparams"))
 #' }
-#' 
+#'
 #' @section Description of the dataframes present in the database:
 #'
 #' \itemize{
