### 2019-04-11 28648b7
fix bug in examples
