+++ man/download_ramlegacy.Rd
@@ -42,8 +42,8 @@ downloading all the older versions (1.0, 2.0, 2.5, 3.0, 4.3) from
 [a github repo](www.github.com/kshtzgupta1/ramlegacy-assets)
 }
 \examples{
-\donttest{
 
+\dontshow{Sys.setenv(RAM_HOME = tempfile())}
 # If version is not specified then current latest version (4.44)
 # will be downloaded
 download_ramlegacy()
@@ -53,7 +53,7 @@ download_ramlegacy(version = "1.0")
 
 # download version 4.40
 download_ramlegacy(version = "4.40")
-}
+
 }
 \seealso{
 Other ramlegacy functions: \code{\link{load_ramlegacy}},

[…]
