+++ cran-comments.md
@@ -1,10 +1,17 @@
 ## Test environments
-* local OS X install, R 3.5.1
-* ubuntu 14.04 (on travis-ci), R 3.5.1
-* win-builder (devel and release)
+* Local - ubuntu 18.04 (R 3.5.1)
+* Travis CI - ubuntu 14.04 (R 3.5.3)
+* Appveyor - Windows Server 2012 R2 x64 (R 3.5.1 (patched))
+* win-builder (oldrelease, release, and devel)
+
+## Note about examples in download_ramlegacy and load_ramlegacy
+The
 
 ## R CMD check results
 
-0 errors | 0 warnings | 1 note
+There were no ERRORs, no WARNINGs no NOTEs
 
-* This is a new release.
+## Downstream dependencies
+
+There are no downstream dependencies
+
