+++ man/load_ramlegacy.Rd
@@ -118,17 +118,17 @@ reference point type along the columns and stocks along the rows
 }
 
 \examples{
-
-\dontshow{Sys.setenv(RAM_HOME = tempfile())}
+\donttest{
+\dontshow{
+Sys.setenv(RAM_HOME = tempfile())
+}
 # first download version 4.44 of the database
 download_ramlegacy(version = "4.44")
-
+}
 
 # get a list containing area and bioparams tables
 # from version 4.44 database
 load_ramlegacy(version = "4.44", tables = c("area", "bioparams"))
-
-
 }
 \seealso{
 Other ramlegacy functions: \code{\link{download_ramlegacy}},
