### 2019-04-11 1059403
add donttest to travis yml
