+++ .travis.yml
@@ -3,6 +3,6 @@
 language: R
 sudo: false
 cache: packages
-
+r_check_args: --run-donttest
 after_success:
   - travis_wait 20 Rscript -e 'library(covr); codecov()'
