### 2019-04-13 fa054aa
fix rappdirs typos
