### 2019-04-13 1534ea4
put load exmple within donttest
