### 2019-04-13 ef4f248
fix rappdirs typo in load doc
