+++ man/ramlegacy-package.Rd
@@ -102,7 +102,7 @@ reference point type along the columns and stocks along the rows
 Ricard, D., Minto, C., Jensen, O.P. and Baum, J.K. (2012)
 Evaluating the knowledge base and status of commercially exploited marine
 species with the RAM Legacy Stock Assessment Database.
-Fish and Fisheries 13 (4) 380-398. <doi:10.1111/j.1467-2979.2011.00435.x.>
+Fish and Fisheries 13 (4) 380-398. <doi:10.1111/j.1467-2979.2011.00435.x>
 }
 \seealso{
 \url{www.ramlegacy.org}
