### 2019-04-13 e3deef5
update docs for new vers lang
