+++ appveyor.yml
@@ -15,6 +15,9 @@ cache:
 
 # Adapt as necessary starting from here
 
+environment:
+r_check_args: --run-donttest
+
 build_script:
   - travis-tool.sh install_deps
 
