### 2019-04-13 5f33666
rm net_check & improve exmples
