### 2019-04-13 a323739
update NEWS.md
