### 2019-04-13 b23ed06
comment out obsolete net_check tests
