### 2019-04-13 d89944a
fix rappdirs link in README
