+++ DESCRIPTION
@@ -6,7 +6,7 @@ Authors@R:
     c(person(given = "Carl",
              family = "Boettiger",
              role = c("aut", "cph"),
-             email = "cboettig@gmail.com", comment="http://orcid.org/0000-0002-1642-628X"),
+             email = "cboettig@gmail.com", comment = c(ORCID = "http://orcid.org/0000-0002-1642-628X")),
      person(given = "Kshitiz",
              family = "Gupta",
              role = c("aut", "cre", "cph"),
@@ -14,11 +14,11 @@ Authors@R:
      person(given = "Sam",
             family = "Albers",
             role = c("rev"),
-            email = "sam.albers@gmail.com", comment = "https://orcid.org/0000-0002-9270-7884"),
+            email = "sam.albers@gmail.com", comment = c(ORCID = "https://orcid.org/0000-0002-9270-7884")),
      person(given = "Jamie",
             family = "Afflerbach",
             role = c("rev"),
-            email = "afflerbach@nceas.ucsb.edu", comment = "https://orcid.org/0000-0002-5215-9342"),     
+            email = "afflerbach@nceas.ucsb.edu", comment = c(ORCID = "https://orcid.org/0000-0002-5215-9342")),     
       person(given = "RAM Legacy Stock Assessment Database",
       role = "dtc"))
 Description: Contains functions to download, cache and read in Excel version of the
@@ -27,7 +27,7 @@ Description: Contains functions to download, cache and read in Excel version of 
     information about the database can be found at <http://ramlegacy.org/>. Ricard, D.,
     Minto, C., Jensen, O.P. and Baum, J.K. (2012) Evaluating the knowledge base and
     status of commercially exploited marine species with the RAM Legacy Stock Assessment
-    Database. Fish and Fisheries 13 (4) 380-398. <doi:10.1111/j.1467-2979.2011.00435.x>
+    Database. Fish and Fisheries 13 (4) 380-398 <doi:10.1111/j.1467-2979.2011.00435.x>.
 License: MIT + file LICENSE
 URL: https://github.com/ropensci/ramlegacy
 BugReports: https://github.com/ropensci/ramlegacy/issues
