### 2019-04-14 5e956df
update NEWS
