+++ NEWS.md
@@ -28,11 +28,11 @@ And constructed with the following guidelines:
 
 v0.1.0
 ------
-
 * Initial Release
 
 v0.2.0
 -----
-* Removed the loading behavior of library(ramlegacy) leaving the loading solely to load_ramlegacy which instead of assigning all the tables to global environment now just returns a list of the tables.
-* Additionally, if the user wants specific tables from a version of the database load_ramlegacy now supports that behavior.
-* We have also modified download_ramlegacy and load_ramlegacy so that while the default is still to download and read in the dataframes from the rappdirs directory the functions now also support downloading to a location chosen by the user and reading from that location.
+* Removed the loading behavior of `library(ramlegacy)` leaving the loading solely to `load_ramlegacy`
+* Instead of assigning all the tables to global environment `load_ramlegacy` now just returns a list of tables that can be specified through `tables` argument.
+* Modified `download_ramlegacy` and `load_ramlegacy` so that that while the default is still to download and read in the dataframes from the rappdirs directory the functions now also support downloading to a location chosen by the user and reading from that location. 
+
