### 2019-04-14 e270315
update NEWS
