### 2019-04-14 678c2b1
update cran-comments
