### 2019-04-28 b3edc8e
fix CRAN typos
