+++ R/ramlegacy-package.R
@@ -4,7 +4,7 @@
 #' from the excel version of the RAM Legacy Stock Assessment Database,
 #' an online compilation of stock assessment results for commercially exploited
 #' marine populations from around the world. More information about the database
-#' can be found at <http://ramlegacy.org/>.
+#' can be found at <https://ramlegacy.org/>.
 #' @keywords internal
 #' @section Description of the dataframes present in the database:
 #'
