### 2019-05-07 f041a79
implement CRAN changes
