### 2019-05-07 41ea9d0
update cran-comments for resubmission
